- Fixed return value of getnbr - Added error handling for register & direct type

+++ include/common/utils/nbr/getnbr.h
@@ -7,4 +7,6 @@
 
 #pragma once
 
-int getnbr(char const *str);
+#include <stdint.h>
+
+int64_t getnbr(char const *str);
